+++ morecantile/models.py
@@ -207,6 +207,11 @@ class TileMatrixSet(BaseModel):
         return f"<TileMatrixSet title='{self.title}' id='{self.id}'>"
 
     @property
+    def geographic_crs(self) -> CRSType:
+        """Return the TMS's geographic CRS."""
+        return self._geographic_crs
+
+    @property
     def rasterio_crs(self):
         """Return rasterio CRS."""
         return to_rasterio_crs(self.crs)